dengue_updater/requirements.txt
requests>=2.31.0
pandas>=2.0.0
pytrends>=4.9.2
openpyxl>=3.1.2
pyyaml>=6.0
websockets>=12.0
